Improve the Installer, Depend only on brew on MacOS, Fix config to auto-provision first time.

+++ utils/appflow.sh
@@ -1,13 +1,51 @@
 #!/bin/bash
 currdir=`pwd`
 
+cd $HOME
 echo "Installing dependencies...";
 echo "#############################";
 DIST=`uname -s`
 if [ $DIST == 'Darwin' ]; then 
-	brew unlink ansible ; brew unlink ansible20 ; brew reinstall ansible --HEAD;
+	if ! brew ls --versions ansible | grep HEAD; then
+		brew unlink ansible ; brew unlink ansible20 ; brew reinstall ansible --HEAD;
+	fi 
+	brew install caskroom/cask/brew-cask
+	if ! brew cask ls --versions virtualbox; then 
+		brew cask install virtualbox	
+	fi 
+	if ! brew cask ls --versions vagrant; then 
+		brew cask install vagrant
+	fi 
+	if ! brew cask ls --versions vagrant-manager; then 
+		brew cask install vagrant-manager
+	fi 
 else 
-	sudo pip install git+git://github.com/ansible/ansible.git;
+	declare -A osInfo;
+	osInfo[/etc/redhat-release]=dnf
+	osInfo[/etc/SuSE-release]=zypp
+	osInfo[/etc/debian_version]=apt-get
+	PKG=
+	for f in ${!osInfo[@]} 
+	do
+		if [[ -f $f ]]; then
+			PKG=${osInfo[$f]}
+		fi
+	done
+	#sudo pip install git+git://github.com/ansible/ansible.git; 
+	sudo $PKG install -y vagrant virtualbox
+	if (($? == 1)); then
+		echo "Error! Check your dependencies! Without vagrant and virtualbox
+you cannot use AppFlow, But it can still be installed.
+Would You Like to continue?"
+		
+		read  -n 1 -p "[y/N]:" vm
+		echo ""
+		case $vm in
+			[Yy]* ) ;;
+			[Nn]* ) exit 1;;
+			* ) echo "Please answer yes or no.";;
+		esac
+	fi
 fi
 
 echo "#############################";
@@ -21,21 +59,28 @@ grep -q -F '192.168.90.3 testing.centos' /etc/hosts || sudo tee -a /etc/hosts<<<
 echo "#############################";
 echo "Configuring default Dirs..."
 echo "#############################";
-mkdir -p ~/Documents/webdev;
-cd ~/Documents/webdev;
+mkdir -p $HOME/Documents/webdev;
+cd $HOME/Documents/webdev;
 
 echo "#############################";
 echo "Cloning AppFlow..."
 echo "#############################";
 git clone git@github.com:ttssdev/appflow.git;
-ln -s ~/Documents/webdev/appflow ~/appflow;
-sudo cp ~/appflow/appflow /usr/local/bin/appflow
+ln -s $HOME/Documents/webdev/appflow $HOME/appflow;
+sudo mkdir -p /usr/local/bin/
+sudo cp $HOME/appflow/appflow /usr/local/bin/appflow
 sudo chmod +x /usr/local/bin/appflow
 
 echo "#############################";
 echo "Preparing the environment..."
 echo "#############################";
-cd ~/Documents/webdev/appflow
+
+mkdir -p $HOME/.appflow/tenant
+ln -s $HOME/Documents/webdev/appflow/examples/appflow-mrrobot $HOME/.appflow/tenant/appflow-mrrobot
+ln -s $HOME/.appflow/tenant/appflow-mrrobot  $HOME/.appflow/tenant/mrrobot
+ln -s $HOME/Documents/webdev/appflow/examples/vault $HOME/.appflow/tenant/vault 
+cp $HOME/Documents/webdev/appflow/config.example $HOME/.appflow/config
+cd $HOME/Documents/webdev/appflow
 make local ask-sudo-pass=true;
 
 echo "Would you like to initialize the Atlantis VM now?"
